refactor(rsc): consolidate RscElement and SuspenseBoundary into main module
- Move RscElement enum definition from types/mod.rs to rsc/mod.rs
- Move SuspenseBoundary struct definition from types/mod.rs to rsc/mod.rs
- Update import paths in flight/parser.rs to reference types from rsc module
- Update import paths in rendering/html/mod.rs to reference types from rsc module
- Simplify types/mod.rs to only expose public re-exports for elements and tree submodules
- Reduce module fragmentation by consolidating core RSC types in the main module namespace
