+++ crates/rari/src/rsc/flight/parser.rs
@@ -1,5 +1,5 @@
 use crate::error::RariError;
-use crate::rsc::types::{RscElement, SuspenseBoundary};
+use crate::rsc::{RscElement, SuspenseBoundary};
 use rustc_hash::{FxHashMap, FxHashSet};
 use serde_json::Value as JsonValue;
